+++ src/fieldtranslators/TextFieldTranslator.php
@@ -8,7 +8,6 @@ use craft\fields\PlainText;
 use craft\fields\BaseRelationField;
 use craft\helpers\ArrayHelper;
 use craft\redactor\Field as Redactor;
-use craft\ckeditor\Field as CKEditor;
 use miranj\autotranslator\Plugin;
 
 /**
@@ -19,7 +18,6 @@ class TextFieldTranslator implements FieldTranslatorInterface
     public const FIELD_TYPES = [
         PlainText::class,
         Redactor::class,
-        CKEditor::class,
     ];
     
     public static function displayName(): string