+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release notes for the Auto Translator Craft CMS plugin.
 ## Unreleased 0.2.0
 
 - Added support for Craft 5.
+- Added support for nested entries in CKEditor fields.
 
 ## Unreleased 0.1.0